Extract shared build-system filesystem helpers into BuildSystemSupport
The three concrete BuildSystem implementations (SPM, Xcode, Bazel) plus
SetupBuilder.swift each open-coded the same `.framework` / `.swift` /
`.o` directory scans. Pull `collectFrameworks`, `collectGeneratedSources`,
and `collectObjectFiles` into a shared `BuildSystemSupport` enum so each
helper is defined once.

Bazel does not pick up any of the shared helpers — its phases use bazel
queries instead of filesystem scans. SPM's `collectGeneratedSources` and
Xcode's `collectGeneratedSources` keep their public test entry points;
they're now thin path-computing shims that delegate to the shared helper.

Skipped from the plan as too divergent to merge:
  - `runBuildSubprocess`: each system has a distinct error-diagnostic
    shape (Bazel's multi-line "command/cwd/stdout/stderr" block is
    load-bearing per the comment about CI run 25243519727).
  - `verifySwiftModule`: 1-2 line existence checks with build-system-
    specific error types (BuildSystemError vs SetupBuilderError).
